chore: refactor animation module
